requirements.txt
discord.py[speed] >=2.0.1, <3
cryptography >=37.0.0, <38.0.0
lz4 >=3.1.10, <5.0.0
typing-extensions >=4.3, <5
